requirements.txt
py-cord
python-dotenv
mcstatus
